+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ### **v0.1.0**:
 
 - Carousel bug fixes
+- Added showcase projects banner in projects page